backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
sqlalchemy==2.0.30
networkx==3.3
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1

# LLM
requests==2.32.3
cohere==5.5.8